UART qui bloquait à cause de int baudrate dans la fonction uart_init, il fallait passer la variable en unsigned long int. Le probleme venait de avr qui considere les int comme 16 bits. Le calcul de 16*9600 sortait de la plage maximale de int16_t

+++ include/uart.h
@@ -1,5 +1,5 @@
 #include <stdlib.h>
 
-void uart_init(int baudrate);
+void uart_init(unsigned long int baudrate);
 char uart_getchar(void);
 void uart_putchar(char character);
